\name{scrmove-package}
\alias{scrmove-package}
\alias{scrmove}
\docType{package}
\title{
\packageTitle{scrmove}
}
\description{
\packageDescription{scrmove}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scrmove}
\packageIndices{scrmove}
  The function \code{\link[scrmove:scrMoveSemi]{scrmove}} fits the
  SCR+movement model using the conditional-on-n approach in which we
  have to compute the probability of not being detected by integrating
  over [y=0|u][u|s][s]. This is done using Monte Carlo integration, sped
  up with RcppArmadillo and OpenMP.
}
\author{
\packageAuthor{scrmove}

Maintainer: \packageMaintainer{scrmove}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
